fix: the admin's python package in the project folder was causing problems when running tests. Merged in one file.
