--- src/project/admin/__init__.py
@@ -1,4 +1,0 @@
-# Using explicit re-exports for linting purposes
-from .base_admin import ModelAdmin as ModelAdmin
-from .base_admin import ModelAdminMixin as ModelAdminMixin
-from .logentry import LogEntryAdmin as LogEntryAdmin
